Polishing...3/n

+++ MonsterChase/Source/Game/GameUtils.h
@@ -19,9 +19,9 @@ private:
 	~GameUtils() {}
 
 public:
-	inline static engine::math::Vec3D GetRandomVec3D(int max_x = 1, int max_y = 1, int max_z = 1) { return engine::math::Vec3D(static_cast<float>(rand() % max_x), static_cast<float>(rand() % max_y), static_cast<float>(rand() % max_z)); }
+	inline static engine::math::Vec3D GetRandomVec3D(size_t max_x = 1, size_t max_y = 1, size_t max_z = 1) { return engine::math::Vec3D(static_cast<float>(rand() % max_x), static_cast<float>(rand() % max_y), static_cast<float>(rand() % max_z)); }
 
-	inline static MoveDirections GetRandomDirection() { return static_cast<MoveDirections>(rand() % static_cast<int>(MoveDirections::kMoveDirectionMax)); }
+	inline static MoveDirections GetRandomDirection() { return static_cast<MoveDirections>(rand() % static_cast<uint8_t>(MoveDirections::kMoveDirectionMax)); }
 
 }; // class GameUtils
